fix some existing issues

+++ src/main/webapp/app/programming/manage/code-editor/file-browser/delete/code-editor-file-browser-delete.ts
@@ -45,6 +45,13 @@ export class CodeEditorFileBrowserDeleteComponent implements OnInit {
      */
     deleteFile() {
         this.isLoading = true;
+
+        // Guard against deleting PROBLEM_STATEMENT files
+        if (this.fileType === FileType.PROBLEM_STATEMENT) {
+            this.closeModal();
+            return;
+        }
+
         // Make sure we have a filename
         if (this.fileNameToDelete) {
             this.repositoryFileService.deleteFile(this.fileNameToDelete).subscribe({
